+++ griptape/engines/rag/modules/__init__.py
@@ -1,0 +1,18 @@
+from .base_module import BaseModule
+from .query.base_query_processor import BaseQueryProcessor
+from .retrieval.base_retriever import BaseRetriever
+from .retrieval.text_retriever import TextRetriever
+from .generation.base_generator import BaseGenerator
+from .generation.prompt_generator import PromptGenerator
+
+__all__ = [
+    "BaseModule",
+
+    "BaseQueryProcessor",
+
+    "BaseRetriever",
+    "TextRetriever",
+
+    "BaseGenerator",
+    "PromptGenerator",
+]